Update README.md
In this the project the goal is to analyze costumers through the use
of RFM(Recency, Frequency, Model) model, Churn model, MBA(Market Basket Analysis) and
Sentiment Analysis. This is done in order to understand the behaviour, the features
of the costumer base. Once the latters are determinated, the  company can carry out specific
strategies in order to improve its revenues.