Update README.md
 Customers analysis using :RFM, Churn, MBA and Sentiment Analysis

In this the project the goal is to analyze customers through the use
of RFM(Recency, Frequency, Model) model, Churn model, MBA(Market Basket Analysis) and
Sentiment Analysis. This is done in order to understand the behaviour, the features
of the customer base. Once the latters are determinated, the  company can carry out specific
strategies in order to improve its revenues.